+++ s2gos-client/src/s2gos_client/cli/output.py
@@ -16,7 +16,7 @@ from s2gos_common.models import (
     ProcessDescription,
     ProcessList,
 )
-from s2gos_common.process.cli.request import ExecutionRequest
+from s2gos_common.process.request import ExecutionRequest
 
 
 class OutputFormat(str, Enum):
